--Empty Message--

+++ blmarket/ArmyTeleportation.cpp
@@ -63,8 +63,8 @@ bool match() {
     v4[1] = mp(xt[0] - xt[2], yt[0] - yt[2]);
     v4[2] = mp(xt[1] - xt[2], yt[1] - yt[2]);
 
-    int xt = v3[0].first;
-    int yt = v3[0].second;
+    LL xt = v3[0].first;
+    LL yt = v3[0].second;
 
     v4[0].first *= v4[2].second;
     v4[1].first *= v4[2].second;
